layers: LunarXchange #145: Fix windows printf format errors
Windows hates %zu. Changed to %lu and cast size_t to unsigned long.

+++ layers/vk_layer_logging.h
@@ -293,8 +293,8 @@ static inline VkBool32 VKAPI log_callback(
 
     print_msg_flags(msgFlags, msg_flags);
 
-    fprintf((FILE *) pUserData, "%s(%s): object: %#" PRIx64 " type: %d location: %zu msgCode: %d: %s\n",
-             pLayerPrefix, msg_flags, srcObject, objType, location, msgCode, pMsg);
+    fprintf((FILE *) pUserData, "%s(%s): object: %#" PRIx64 " type: %d location: %lu msgCode: %d: %s\n",
+             pLayerPrefix, msg_flags, srcObject, objType, (unsigned long)location, msgCode, pMsg);
     fflush((FILE *) pUserData);
 
     return false;
